backend/requirements.txt
flask
flask-cors
torch
torchvision
gunicorn
pillow
